Fix CandidateDataManagerTest + test autoAddToUpNextSetting

+++ Modules/DataModel/Tests/PocketCastsDataModelTests/AutoAddCandidatesDataManagerTests.swift
@@ -21,22 +21,25 @@ final class AutoAddCandidatesDataManagerTests: XCTestCase {
 
     private func setupDatabase() throws -> DataManager {
         let dbPath = (DataManager.pathToDbFolder() as NSString).appendingPathComponent("podcast_testDB.sqlite3")
-        try FileManager.default.removeItem(atPath: dbPath)
+        if FileManager.default.fileExists(atPath: dbPath) {
+            try FileManager.default.removeItem(atPath: dbPath)
+        }
         let flags = SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE | SQLITE_OPEN_FILEPROTECTION_NONE
         let dbQueue = try XCTUnwrap(FMDatabaseQueue(path: dbPath, flags: flags))
         return DataManager(dbQueue: dbQueue)
     }
 
+    /// Tests new query and autoAddToUpNext property for UpNext candidates
     func testSyncableUpNextSetting() throws {
         try override(flag: .settingsSync, value: true)
 
         let dataManager = try setupDatabase()
+        let newUpNextSetting = AutoAddToUpNextSetting.addFirst
 
         let podcast = Podcast()
         podcast.uuid = "1234"
         podcast.addedDate = Date()
-        podcast.settings.addToUpNext = true
-        podcast.settings.addToUpNextPosition = .top
+        podcast.setAutoAddToUpNext(setting: newUpNextSetting)
 
         let episode = Episode()
         episode.uuid = "1234"
@@ -49,21 +52,24 @@ final class AutoAddCandidatesDataManagerTests: XCTestCase {
 
         let candidates = dataManager.autoAddCandidates.candidates()
 
-        XCTAssertTrue(candidates.contains(where: { $0.episodeUuid == episode.uuid }), "Episode should appear in Up Next candidates")
+        let matchingCandidate = candidates.first(where: { $0.episodeUuid == episode.uuid })
+        XCTAssertNotNil(matchingCandidate, "Episode should appear in Up Next candidates")
+        XCTAssertEqual(matchingCandidate?.autoAddToUpNextSetting, newUpNextSetting)
 
         try reset(flag: .settingsSync)
     }
 
+    /// Tests old query and autoAddToUpNext property for UpNext candidates
     func testOldUpNextSetting() throws {
         try override(flag: .settingsSync, value: false)
 
         let dataManager = try setupDatabase()
+        let newUpNextSetting = AutoAddToUpNextSetting.addFirst
 
         let podcast = Podcast()
         podcast.uuid = "1234"
         podcast.addedDate = Date()
-        podcast.settings.addToUpNext = true
-        podcast.settings.addToUpNextPosition = .top
+        podcast.setAutoAddToUpNext(setting: newUpNextSetting)
 
         let episode = Episode()
         episode.uuid = "1234"
@@ -76,7 +82,9 @@ final class AutoAddCandidatesDataManagerTests: XCTestCase {
 
         let candidates = dataManager.autoAddCandidates.candidates()
 
-        XCTAssertTrue(candidates.contains(where: { $0.episodeUuid == episode.uuid }), "Episode should appear in Up Next candidates")
+        let matchingCandidate = candidates.first(where: { $0.episodeUuid == episode.uuid })
+        XCTAssertNotNil(matchingCandidate, "Episode should appear in Up Next candidates")
+        XCTAssertEqual(matchingCandidate?.autoAddToUpNextSetting, newUpNextSetting)
 
         try reset(flag: .settingsSync)
     }
